microcloud/cmd/microcloud/main/init: Retry disk selection on error
This adds an askRetry helper to the CLI, which will re-run the given
function if an error occurs. This is helpful for disk selection in
particular because there are constraints that might not be met in
regards to which disks can be selected on which peers. So allowing a
retry rather than just failing will be helpful.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>