microcloud/service/service/handler: Add ServiceExists helper
The previous method of creating a whole struct just to stat the state
directory was rather inefficient, so this makes it a bit dumber, but
also a bit clearer.

Additionally, on init, MicroCloud will now ask for confirmation before
going ahead with the cluster formation if a service is missing.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>