microcloud/mdns: Broadcast over all interfaces
This adds a ServiceSize constant to the mdns broadcast specifying how
many simultaneous variations of the existing mdns service we are
broadcasting.

On Lookup, we must specify a service key for the broadcast we want,
but when broadcasting over every interface, these records overlap and
prevent us from finding all of them, as we keep picking up the last
one we found. Assigning explicit keys to each broadcast is difficult
as we  won't know how many records are available. So instead, we will
send up  to `ServiceSize` variations of the service
(_microcloud_0, _microcloud_1, microcloud_2... etc) and we will only
double up on the same service key if we have more than `ServiceSize`
interfaces available.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>