microcloud/cmd/microcloud/main/init: Create InitSystem type
InitSystem will represent all the init configuration collected by the
interactive CLI for each peer. A map of InitSystem will be passed around
the CLI functions and slowly added to, before it is finally used to
initialize the cluster.

This will hopefully make debugging MicroCloud much easier as we will
have a snapshot of all collected information throughout the init
process. It might also facilitate some testing of the CLI.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>