microcloud/cmd/microcloud: Use Default to pass table data to template
Default is an interface{} so we can abuse this to pass the border and
header strings into the template without having to regenerate it.

Since we are co-opting the `Default` field to pass in template data,
this commit adds a new template function which effectively runs the
the functionality defined here:

https://github.com/go-survey/survey/blob/fa37277e6394c29db7bcc94062cb30cd7785a126/multiselect.go#L235-L257

but on the `DefaultRows` field of the templateData.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>