microcloud/service: Add tokens endpoint
During init, to handle the case where another system is already
clustered on a particular service, we need to be able to request
this node to issue a token for us from an as-yet untrusted system.
This endpoint is untrusted by the cluster, but authenticated with a
secret generated during mDNS lookup, so we can use it as a proxy to
the unix socket on the remote system, where we will be trusted and
can issue a token.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>