cmd/microcloud: Reorganize network question helper
It was a bit weird that we generate a FAN network and then overwrite it.
Especially now that we need to check existing cluster members,
this is a bit more expensive.

So this refactors askNetwork into two helpers:
askOVNNetwork which sets up the OVN network, then if no network is
created by that helper, we fallback to the FAN implementation.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>