microcloud/service: Use same remoteClient logic for all services
remoteClient gets a client to underlying service via the MicroCloud
proxy, usually used for communication between microcloud-eligibile
systems during init. They are normally authenticated with a secret
obtained by mDNS and set in the request header.

It should also be allowed for trusted systems (cluster members) to
communicate over this proxy, but the previous client implementation
overwrote the TLS config of the client, so this now preserves it.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>