microcloud/cmd/microcloud: Don't fully overwrite the default profile
Rather than overwriting the default profile when adding devices, this
appends the devices and config to the default profile. When adding a new
service to microcloud that was skipped during initial setup, we may have
to add a new device to the default profile but we shouldn't remove
anything we added during first init.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>