microcloud/service: Only set source.wipe if user explicitly agrees
If LXD doesn't have the corresponding schema extension for the
source.wipe config key, MicroCloud will never ask the user if they want
to wipe disks, but will still populate the config with a 'false' value,
causing an error.

Instead, we can simply omit the key if it's false, and it will be
implicitly considered false.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>