microcloud/service/microceph: Add the `ClusterConfig` function
When a Ceph cluster is deployed, we need to be able to query its global configuration.
For example, we'll need this client logic to know if a deployed Ceph cluster has an existing
`cluster_network` parameter. This will be required for adding a new cluster member using the
dedicated Ceph subnet.

Signed-off-by: Gabriel Mougard <gabriel.mougard@canonical.com>