client: Add UseAuthProxy client function
This applies a proxy to a microcluster client for a different service,
which instead sends the request to be handled by the MicroCloud daemon,
which then forwards it to the unix socket of the corresponding service.

If a secret is given, it sets `InsecureSkipVerify` so that we can skip
TLS in favour of the pre-cluster-formation auth secret in the request
header.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>