cmd/microcloud: Explicitly set bootstrap status
Relying on the system map having the local entry to determine if we need
to bootstrap is not always accurate. So instead we can just set what we
expect given the command we are running.

The exception is the preseed case where the system map is not modified
from what is supplied by the preseed.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>