cmd/microcloud: Check if existing cluster has storage pools
If the existing cluster does not already have storage pools, we should
include its disks in the table.

First, we fetch the disks for all cluster members if a pool that we want
doesn't exist.

Then we create a new InitSystem map with both the joiners and the
existing cluster members.

Then, for ZFS and Ceph setup, we update the InitSystem entries with the
new disks and reapply them to the temporary map.

Finally, the map is reconciled with the global InitSystem map.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>