cmd/microcloud: Handle default find_min
Since many filters can be defined, selecting a proper default value for
find_min is difficult.

For local storage, it's a bit easier because later on we validate that
at least all systems must have an associated disk. So we can easily
default to 1 in this case.

For distributed storage, not all systems need to have a disk, and there
are also non-recommended non-HA setups with fewer than 3 systems
specifying a disk. In this case, we will require each filter to define a
minimum `find_min` value of at least 1.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>