test/suites: Rework combinations test
Many of the cases of this test are actually redundant, and since the
behaviour between <3 and >3 members is no longer different, we can drop
most of those cases and just test using 2 systems.

This should make this test way faster, though it's still likely to break
with 4 systems.

Additionally, rather than looping, since the cases are fewer and less
ordered than before, it's actually simpler to just list the
combinations.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>