mdns: Lookup only a single system and increase the version number
This reworks the mDNS code to support looking up only a single system over the configured
lookup interface which matches the one used by the local MicroCloud's listen address.
It has the added effect that we can also get rid of the Interface field in the ServerInfo struct.

In addition the broadcast version number is increased to prevent collissions with earlier versions of MicroCloud.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>