test/includes/microcloud: Fix bug with killing too many processes
When cleaning up MicroOVN and MicroCeph left over processes, the 'sh' invocation itself also has to
be filtered out of the list.
Otherwise 'kill -9' will also kill the other half of 'lxc exec' inside the instance as the process spawned
by the lxd-agent starts with 'sh -c ..micro*...' which matches the patterns.

This led to problems during test execution as the 'lxc exec' command returned 255 when resetting MicroOVN.
I have applied the same pattern to the MicroCeph cleanup even though the error was only observed for MicroOVN.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>