multicast: Add StopResponder func to allow waiting for a closed connection
In the test suite we found that subsequent starts of the multicast discovery responder might conflict
with each other as the cancelled context might not have yet released the connection from the previous test run.

By having a dedicated StopResponder that closes the active connection we can ensure that the responders
connection is closed when the func returns.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>