cmd/microcloud/selection_table: Wait until stdio is set in the go routine
We recently saw panics occuring in the pipeline when in TEST_CONSOLE=1 mode.
This is caused by the Render function returning before the pseudo terminals stdio
is configured for the survey which led to panics down the line in case
stdout wasn't yet set.

A fast call to table.Update right after table.Render is causing this.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>