service: Untangle MicroOVN from client functions
This allows updating MicroCloud's version of Microcluster as we can
always infer our preferred version of Microcluster and don't have
to supply it to MicroOVN's client functions which might require
a different version of Microcluster.

Future changes to the client functions should make use of API extension
checks (same as in the LXD client).

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>