test/includes/microcloud: Hold all snaps to prevent unplanned refresh
The test suite should hold the lxd, microceph, microovn and microcloud snaps to not cause
unplanned refreshes during test execution.

This was observed in the pipeline every now and then and caused some of the daemons to be stopped out of nowhere
which resulted in API endpoints returning EOF.
In the daemon logs this is clearly indicated by systemd stopping the respective service and starting it afterwards.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>