cmd/microcloud: Protect the selection of partitions
When selecting a partition, its parent disk cannot be selected too when configuring remote storage.

Also when selecting a disk with partitions, offer to wipe the disk to remove the partitions.
In all other cases wiping isn't necessary as the disk is only shown in case it doesn't contain any FS.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>