test/suites/upgrade: wait for instances to be ready before exec'ing into them
This will allow time for `v1` to be booted which can be long for VMs. It
probably never worked but we didn't notice as our CI has `SKIP_VM_LAUNCH=1`.
This was not tested, just opportunistically "fixed".

It seems that we were lucky enough for `c1` to always be ready for `lxc
exec`'ing into it right after launch but it's best to play it safe even for
containers.

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>

test/suites/upgrade: wait for instances to be ready before exec'ing into them

Signed-off-by: Simon Deziel <simon.deziel@canonical.com>