cmd/microcloud: Exit add command in case no new system got selected
The check itself cannot be done inside the initiating session because the same session func is used
for both the init and add subcommands.
In case no new system got selected with the init subcommand, the questionnaire falls back to creating a single node MicroCloud.
In case of the add subcommand there has to be at least one new system to continue.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>