cmd/microcloud: Require either IPv4 or IPv6 gateway on uplink
Not configuring the gateway on the uplink configures an OVN network without external connectivity.
Whilst that might be a valid use case, currently this is not really supported in LXD as you have to provide
an uplink network even though it's not used.

Until this is supported in LXD, we simply don't allow this type of configuration.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>