cmd/microcloud/ask: Collect list of already existing remote storage disks
This allows understanding which disks on which cluster members are already configured for remote storage.
The information can then be yielded to the user and allows skipping the selection of additional disks
in case the user only want to configure distributed storage without adding additional disks to MicroCeph.

That scenario is important when adding an existing MicroCeph cluster to MicroCloud.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>