cmd/microcloud/ask: Collect list of already existing remote storage disks
This allows understanding which disks on which cluster members are already configured for remote storage.
The information can then be yielded to the user and allows skipping the selection of additional disks
in case the user only wants to configure distributed storage without adding additional disks to MicroCeph.

That scenario is important when adding an existing MicroCeph cluster to MicroCloud.

Also move the availableDiskCount to narrow its scope to where it is actually used.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit f59537733433f4e507fe466ae8afe0a28c3cc9cc)