cmd/microcloud: Include both networks and storage pools when asking for profile replacement
Usually the askProfileUpdate helper should ask in case either the default profile's network or storage pool (root vol) is about to be changed.
However when moving the logic in the the askProfileUpdate helper, only the loop which was populating the profile's networks was moved
(See commit https://github.com/canonical/microcloud/pull/367/commits/29764bcd6a4e54819a2a575f657970cfb7cc7f7a).

The storage pools where not moved which caused the helper to only ask for confirmation if the network got changed.
But the storage pool was overwritten in any case.
If a cluster was formed only using local storage and MicroCeph got added at a later point in time, the profile's storage pool
was overwritten to use the new remote storage pool without asking for confirmation.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>