cmd/microcloud: Allow having target pools and join config at the same time
When differentiating between both then a system can either setup target pools or have ones to join.
However this doesn't work in a situation where a system is both joining a storage pool (e.g. local) and
is initiating the creation of a new storage (e.g. distributed storage) which is possible when adding new systems.

In this case both the join config and target pools are populated so we should consider both lists for the final
list of storage pools.

Otherwise the added systems won't have any of the storage.* settings for backups and image vols.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit c5550546951ce04819b9790970d2ccc9093f1b4e)