test/suites/e2e: Reduce the number of evacuation counts
As we have found the root cause for the race (LXD network/storage waitready), we don't anymore have to
retry the evacuation then times.

However until everything stabilizes it still makes sense to run it more than one time.
We can always reduce the number again to one in the future.
But running it three times should be a good compromise between length of execution and test coverage.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>