github: Run scheduled tests fully also for doc only changes
In one of the last scheduled weekly test runs I have observed that in case the last commit on main
consists only of doc changes, it doesn't run the full test suite:
https://github.com/canonical/microcloud/actions/runs/16545336169
This is due to the fact that we run only the resource intensive code and system tests for changes that actually
affect this part of the code.

By making the check for except_docs dependent on the github.event_name, we can ensure that the code (and system) tests
will always run regardless whether or not the last commit on main contains only doc changes.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>
(cherry picked from commit 86f9284e16a4e797b3b3a9ab94d31e73000c3a4f)