cmd/microcloud: Move fingerprint verification info post passphrase input
This is a request raised in https://github.com/canonical/microcloud/issues/949.

When printing the fingerprint verification info before entering the passphrase, we are telling the user to perform
an action even though this can only be done after entering the passphrase and the joiner reached out to the initator.

Signed-off-by: Julian Pelizäus <julian.pelizaeus@canonical.com>